Refactor move `DataTask` to `steps/data.py`
